packed JBSplits fully by increasing percent fidelity and projectId fidelity

+++ contracts/structs/JBSplit.sol
@@ -7,8 +7,8 @@ struct JBSplit {
   // A flag that only has effect if a projectId is also specified, and that project has issued its tokens.
   // If so, this flag indicates if the tokens that result from making a payment to the project should be delivered staked or unstaked to the beneficiary.
   bool preferClaimed;
-  // The percent of the whole group that this split occupies. This number is out of 10000000.
-  uint24 percent;
+  // The percent of the whole group that this split occupies. This number is out of 1000000000.
+  uint32 percent;
   // Specifies if the split should be unchangeable until the specifies time comes, with the exception of extending the lockedUntil period.
   uint48 lockedUntil;
   // The role the  beneficary depends on whether or not projectId is specified, or whether or not allocator is specified.
@@ -20,5 +20,5 @@ struct JBSplit {
   IJBSplitAllocator allocator;
   // If an allocator is not set but a projectId is set, funds will be sent to the Juicebox treasury belonging to the project who's ID is specified.
   // Resulting tokens will be routed to the beneficiary with the unstaked token prerence respected.
-  uint56 projectId;
+  uint88 projectId;
 }
